requirements.txt
requests>=2.28.0
pydantic>=2.0.0
pandas>=1.5.0
numpy>=1.23.0
python-dotenv>=1.0.0
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.0
asyncpg>=0.24.0
pytest>=6.2.5
pytest-asyncio>=0.16.0
pytest-cov>=2.12.0
httpx>=0.23.0
aiosqlite>=0.17.0
